# Proposal to opmize "or" regex on the FastRegexMatcher