@alanprot 2024-02-24
@bboreham @bwplotka @jesusvazquez 🥺 thoughts?