@bboreham 2024-02-24
See also https://github.com/prometheus/prometheus/pull/13461