@alanprot 2024-02-26
@bboreham yeah.. maybe https://github.com/prometheus/prometheus/pull/13461 also handle this case, we maybe can wait for that PR and i can after check if this case is handled there.